Abstract MultiUploader In order to have a customizable list rendering, create an abstract MultipleUploader taking a type extending IsWidget and ForWidgets which takes care of the display part
